moved shared constants to constants file
